authorizations implemented

+++ src/api/playlists/handler.js
@@ -12,8 +12,8 @@ class PlaylistsHandler {
   async postPlaylistHandler (request, h) {
     this._validator.validatePostPlaylistPayload(request.payload)
 
-    const { name } = request.payload
     const { id: owner } = request.auth.credentials
+    const { name } = request.payload
 
     const playlistId = await this._playlistsService.addPlaylist(name, owner)
 
@@ -28,7 +28,8 @@ class PlaylistsHandler {
   }
 
   async getPlaylistsHandler (request, h) {
-    const playlists = await this._playlistsService.getPlaylists()
+    const { id: credentialId } = request.auth.credentials
+    const playlists = await this._playlistsService.getPlaylists(credentialId)
     return h.response({
       status: 'success',
       data: { playlists }
@@ -37,7 +38,9 @@ class PlaylistsHandler {
 
   async deletePlaylistHandler (request, h) {
     const { id } = request.params
+    const { id: credentialId } = request.auth.credentials
 
+    await this._playlistsService.verifyPlaylistOwner(id, credentialId)
     await this._playlistsService.deletePlaylistById(id)
 
     return h.response({
@@ -48,21 +51,30 @@ class PlaylistsHandler {
 
   async postSongToPlaylistHandler (request, h) {
     this._validator.validatePostPlaylistSongPayload(request.payload)
+    const { id: credentialId } = request.auth.credentials
 
     const { songId } = request.payload
     await this._songsService.getSongById(songId)
 
     const { id: playlistId } = request.params
+
+    await this._playlistsService.verifyPlaylistOwner(playlistId, credentialId)
     await this._playlistsService.addSongToPlaylist(playlistId, songId)
 
-    return h.response({
+    const response = h.response({
       status: 'success',
       message: 'Berhasil menambahkan lagu kedalam playlist'
     })
+
+    response.code(201)
+    return response
   }
 
   async getSongsInPlaylistHandler (request, h) {
     const { id } = request.params
+    const { id: credentialId } = request.auth.credentials
+
+    await this._playlistsService.verifyPlaylistOwner(id, credentialId)
     const { playlist, songs } = await this._playlistsService.getSongsInPlaylist(id)
 
     return h.response({
@@ -73,11 +85,13 @@ class PlaylistsHandler {
 
   async deleteSongInPlaylistHandler (request, h) {
     this._validator.validateDeletePlaylistSongPayload(request.payload)
+    const { id: credentialId } = request.auth.credentials
 
     const { songId } = request.payload
     await this._songsService.getSongById(songId)
 
     const { id: playlistId } = request.params
+    await this._playlistsService.verifyPlaylistOwner(playlistId, credentialId)
     await this._playlistsService.deleteSongInPlaylist(playlistId, songId)
 
     return h.response({
